Split contract store and contract gateway
