+++ internal/contracts/rest2eth_test.go
@@ -75,7 +75,7 @@ func (m *mockREST2EthDispatcher) DispatchDeployContractSync(ctx context.Context,
 	}
 }
 
-type mockABILoader struct {
+type mockContractResolver struct {
 	loadABIError           error
 	deployMsg              *messages.DeployContract
 	abiInfo                *abiInfo
@@ -84,36 +84,37 @@ type mockABILoader struct {
 	resolveContractErr     error
 	nameAvailableError     error
 	capturedAddr           string
-	postDeployError        error
 }
 
-func (m *mockABILoader) SendReply(message interface{}) {
-
-}
-
-func (m *mockABILoader) lookupContractInstance(addrHex string) (*contractInfo, error) {
+func (m *mockContractResolver) lookupContractInstance(addrHex string) (*contractInfo, error) {
 	m.capturedAddr = addrHex
 	return m.contractInfo, m.loadABIError
 }
 
-func (m *mockABILoader) resolveContractAddr(registeredName string) (string, error) {
+func (m *mockContractResolver) resolveContractAddr(registeredName string) (string, error) {
 	return m.registeredContractAddr, m.resolveContractErr
 }
 
-func (m *mockABILoader) loadDeployMsgByID(addrHex string) (*messages.DeployContract, *abiInfo, error) {
+func (m *mockContractResolver) loadDeployMsgByID(addrHex string) (*messages.DeployContract, *abiInfo, error) {
 	return m.deployMsg, m.abiInfo, m.loadABIError
 }
 
-func (m *mockABILoader) checkNameAvailable(name string, isRemote bool) error {
+func (m *mockContractResolver) checkNameAvailable(name string, isRemote bool) error {
 	return m.nameAvailableError
 }
 
-func (m *mockABILoader) PreDeploy(msg *messages.DeployContract) error { return nil }
-func (m *mockABILoader) PostDeploy(msg *messages.TransactionReceipt) error {
+type mockGateway struct {
+	postDeployError error
+}
+
+func (m *mockGateway) SendReply(message interface{}) {
+}
+func (m *mockGateway) PreDeploy(msg *messages.DeployContract) error { return nil }
+func (m *mockGateway) PostDeploy(msg *messages.TransactionReceipt) error {
 	return m.postDeployError
 }
-func (m *mockABILoader) AddRoutes(router *httprouter.Router) { return }
-func (m *mockABILoader) Shutdown()                           { return }
+func (m *mockGateway) AddRoutes(router *httprouter.Router) { return }
+func (m *mockGateway) Shutdown()                           { return }
 
 type mockRPC struct {
 	capturedMethod string
@@ -218,22 +219,39 @@ func newTestDeployMsg(t *testing.T, addr string) *remoteregistry.DeployContractW
 func newTestREST2Eth(t *testing.T, dispatcher *mockREST2EthDispatcher) (*rest2eth, *mockRPC, *httprouter.Router) {
 	mockRPC := &mockRPC{}
 	deployMsg := newTestDeployMsg(t, "")
-	abiLoader := &mockABILoader{
+	gateway := &mockGateway{}
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg:    &deployMsg.DeployContract,
 	}
 	mockProcessor := &mockProcessor{}
-	r := newREST2eth(abiLoader, mockRPC, nil, nil, mockProcessor, dispatcher, dispatcher)
+	r := newREST2eth(gateway, contractResolver, mockRPC, nil, nil, mockProcessor, dispatcher, dispatcher)
 	router := &httprouter.Router{}
 	r.addRoutes(router)
 
 	return r, mockRPC, router
 }
 
-func newTestREST2EthCustomAbiLoader(dispatcher *mockREST2EthDispatcher, abiLoader *mockABILoader) (*rest2eth, *mockRPC, *httprouter.Router) {
+func newTestREST2EthCustomGateway(t *testing.T, dispatcher *mockREST2EthDispatcher, gateway *mockGateway) (*rest2eth, *mockRPC, *httprouter.Router) {
 	mockRPC := &mockRPC{}
+	deployMsg := newTestDeployMsg(t, "")
+	contractResolver := &mockContractResolver{
+		contractInfo: &contractInfo{},
+		deployMsg:    &deployMsg.DeployContract,
+	}
 	mockProcessor := &mockProcessor{}
-	r := newREST2eth(abiLoader, mockRPC, nil, nil, mockProcessor, dispatcher, dispatcher)
+	r := newREST2eth(gateway, contractResolver, mockRPC, nil, nil, mockProcessor, dispatcher, dispatcher)
+	router := &httprouter.Router{}
+	r.addRoutes(router)
+
+	return r, mockRPC, router
+}
+
+func newTestREST2EthCustomContractResolver(dispatcher *mockREST2EthDispatcher, contractResolver *mockContractResolver) (*rest2eth, *mockRPC, *httprouter.Router) {
+	mockRPC := &mockRPC{}
+	gateway := &mockGateway{}
+	mockProcessor := &mockProcessor{}
+	r := newREST2eth(gateway, contractResolver, mockRPC, nil, nil, mockProcessor, dispatcher, dispatcher)
 	router := &httprouter.Router{}
 	r.addRoutes(router)
 
@@ -252,10 +270,7 @@ func newTestREST2EthAndMsg(t *testing.T, dispatcher *mockREST2EthDispatcher, fro
 }
 
 func newTestREST2EthAndMsgPostDeployError(t *testing.T, dispatcher *mockREST2EthDispatcher, from, to string, bodyMap map[string]interface{}) (*rest2eth, *mockRPC, *httprouter.Router, *httptest.ResponseRecorder, *http.Request) {
-	deployMsg := newTestDeployMsg(t, "")
-	abiLoader := &mockABILoader{
-		contractInfo:    &contractInfo{},
-		deployMsg:       &deployMsg.DeployContract,
+	gateway := &mockGateway{
 		postDeployError: fmt.Errorf("pop"),
 	}
 	body, _ := json.Marshal(&bodyMap)
@@ -263,7 +278,7 @@ func newTestREST2EthAndMsgPostDeployError(t *testing.T, dispatcher *mockREST2Eth
 	req.Header.Add("x-firefly-from", from)
 	res := httptest.NewRecorder()
 
-	r, mockRPC, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	r, mockRPC, router := newTestREST2EthCustomGateway(t, dispatcher, gateway)
 
 	return r, mockRPC, router, res, req
 }
@@ -387,8 +402,8 @@ func TestDeployContractAsyncDuplicate(t *testing.T) {
 		},
 	}
 	r, _, router, res, _ := newTestREST2EthAndMsg(t, dispatcher, from, "", bodyMap)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.nameAvailableError = fmt.Errorf("spent already")
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.nameAvailableError = fmt.Errorf("spent already")
 	body, _ := json.Marshal(&bodyMap)
 	req := httptest.NewRequest("POST", "/abis/abi1?fly-privateFrom=0xdC416B907857Fa8c0e0d55ec21766Ee3546D5f90&fly-privateFor=0xE7E32f0d5A2D55B2aD27E0C2d663807F28f7c745&fly-privateFor=0xB92F8CebA52fFb5F08f870bd355B1d32f0fd9f7C", bytes.NewReader(body))
 	req.Header.Add("x-firefly-from", from)
@@ -844,8 +859,8 @@ func TestSendTransactionUnknownRegisteredName(t *testing.T) {
 	from := "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8"
 	dispatcher := &mockREST2EthDispatcher{}
 	r, _, router, res, req := newTestREST2EthAndMsg(t, dispatcher, from, to, bodyMap)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.resolveContractErr = fmt.Errorf("unregistered")
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.resolveContractErr = fmt.Errorf("unregistered")
 	router.ServeHTTP(res, req)
 
 	assert.Equal(404, res.Result().StatusCode)
@@ -888,7 +903,7 @@ func TestSendTransactionBadMethodABI(t *testing.T) {
 			Request: "request1",
 		},
 	}
-	abiLoader := &mockABILoader{
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg: &messages.DeployContract{
 			ABI: ethbinding.ABIMarshaling{
@@ -900,7 +915,7 @@ func TestSendTransactionBadMethodABI(t *testing.T) {
 			},
 		},
 	}
-	_, _, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	_, _, router := newTestREST2EthCustomContractResolver(dispatcher, contractResolver)
 	req := httptest.NewRequest("GET", "/contracts/0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8/badmethod", bytes.NewReader([]byte{}))
 	req.Header.Add("x-firefly-from", "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8")
 	res := httptest.NewRecorder()
@@ -922,7 +937,7 @@ func TestSendTransactionBadEventABI(t *testing.T) {
 			Request: "request1",
 		},
 	}
-	abiLoader := &mockABILoader{
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg: &messages.DeployContract{
 			ABI: ethbinding.ABIMarshaling{
@@ -934,7 +949,7 @@ func TestSendTransactionBadEventABI(t *testing.T) {
 			},
 		},
 	}
-	_, _, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	_, _, router := newTestREST2EthCustomContractResolver(dispatcher, contractResolver)
 	req := httptest.NewRequest("POST", "/contracts/0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8/badevent/subscribe", bytes.NewReader([]byte{}))
 	req.Header.Add("x-firefly-from", "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8")
 	res := httptest.NewRecorder()
@@ -956,7 +971,7 @@ func TestSendTransactionBadConstructorABI(t *testing.T) {
 			Request: "request1",
 		},
 	}
-	abiLoader := &mockABILoader{
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg: &messages.DeployContract{
 			ABI: ethbinding.ABIMarshaling{
@@ -968,7 +983,7 @@ func TestSendTransactionBadConstructorABI(t *testing.T) {
 			},
 		},
 	}
-	_, _, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	_, _, router := newTestREST2EthCustomContractResolver(dispatcher, contractResolver)
 	req := httptest.NewRequest("POST", "/abis/testabi", bytes.NewReader([]byte{}))
 	req.Header.Add("x-firefly-from", "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8")
 	res := httptest.NewRecorder()
@@ -990,13 +1005,13 @@ func TestSendTransactionDefaultConstructorABI(t *testing.T) {
 			Request: "request1",
 		},
 	}
-	abiLoader := &mockABILoader{
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg: &messages.DeployContract{
 			ABI: ethbinding.ABIMarshaling{}, // completely empty ABI is ok
 		},
 	}
-	_, _, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	_, _, router := newTestREST2EthCustomContractResolver(dispatcher, contractResolver)
 	req := httptest.NewRequest("POST", "/abis/testabi", bytes.NewReader([]byte{}))
 	req.Header.Add("x-firefly-from", "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8")
 	res := httptest.NewRecorder()
@@ -1014,7 +1029,7 @@ func TestSendTransactionUnnamedParamsABI(t *testing.T) {
 			Request: "request1",
 		},
 	}
-	abiLoader := &mockABILoader{
+	contractResolver := &mockContractResolver{
 		contractInfo: &contractInfo{},
 		deployMsg: &messages.DeployContract{
 			ABI: ethbinding.ABIMarshaling{
@@ -1027,7 +1042,7 @@ func TestSendTransactionUnnamedParamsABI(t *testing.T) {
 			},
 		},
 	}
-	_, _, router := newTestREST2EthCustomAbiLoader(dispatcher, abiLoader)
+	_, _, router := newTestREST2EthCustomContractResolver(dispatcher, contractResolver)
 	req := httptest.NewRequest("POST", "/abis/testabi/0x29fb3f4f7cc82a1456903a506e88cdd63b1d74e8/unnamedparamsmethod", bytes.NewReader([]byte{}))
 	req.Header.Add("x-firefly-from", "0x66c5fe653e7a9ebb628a6d40f0452d1e358baee8")
 	q := req.URL.Query()
@@ -1078,8 +1093,8 @@ func TestSendTransactionInvalidContract(t *testing.T) {
 		},
 	}
 	r, _, router, res, req := newTestREST2EthAndMsg(t, dispatcher, from, to, bodyMap)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.loadABIError = fmt.Errorf("pop")
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.loadABIError = fmt.Errorf("pop")
 	router.ServeHTTP(res, req)
 
 	assert.Equal(404, res.Result().StatusCode)
@@ -1108,8 +1123,8 @@ func TestDeployContractInvalidABI(t *testing.T) {
 	body, _ := json.Marshal(&bodyMap)
 	req := httptest.NewRequest("POST", "/abis/abi1?fly-sync", bytes.NewReader(body))
 	req.Header.Add("x-firefly-from", from)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.loadABIError = fmt.Errorf("pop")
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.loadABIError = fmt.Errorf("pop")
 	router.ServeHTTP(res, req)
 
 	assert.Equal(404, res.Result().StatusCode)
@@ -1183,13 +1198,13 @@ func TestSendTransactionRegisteredName(t *testing.T) {
 		},
 	}
 	r, _, router, res, _ := newTestREST2EthAndMsg(t, dispatcher, from, to, bodyMap)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.registeredContractAddr = "c6c572a18d31ff36d661d680c0060307e038dc47"
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.registeredContractAddr = "c6c572a18d31ff36d661d680c0060307e038dc47"
 	req := httptest.NewRequest("POST", "/contracts/"+to+"/set?i=999&s=msg&fly-ethvalue=12345", bytes.NewReader([]byte("{}")))
 	req.Header.Set("x-firefly-from", from)
 	router.ServeHTTP(res, req)
 
-	assert.Equal("c6c572a18d31ff36d661d680c0060307e038dc47", abiLoader.capturedAddr)
+	assert.Equal("c6c572a18d31ff36d661d680c0060307e038dc47", contractResolver.capturedAddr)
 	assert.Equal(202, res.Result().StatusCode)
 }
 func TestSendTransactionMissingParam(t *testing.T) {
@@ -1554,8 +1569,8 @@ func TestSubscribeWithAddressBadAddress(t *testing.T) {
 
 	dispatcher := &mockREST2EthDispatcher{}
 	r, _, router := newTestREST2Eth(t, dispatcher)
-	abiLoader := r.gw.(*mockABILoader)
-	abiLoader.resolveContractErr = fmt.Errorf("unregistered")
+	contractResolver := r.cr.(*mockContractResolver)
+	contractResolver.resolveContractErr = fmt.Errorf("unregistered")
 	r.subMgr = &mockSubMgr{
 		sub: &events.SubscriptionInfo{ID: "sub1"},
 	}
